# JsonParamBinderAnalyzer optimizations