With ImmutabilityAnalyzer as reference (given run variability):

| Analyzer | Before | After |
| -- | -- | -- |
| ImmutabilityAnalyzer | 11.804s | 11.681s |
| JsonParamBinderAnalyzer | 5.418s | 2.227s |